feat: propagate Subscriber deletion to Device Groups (#299)
This PR propagates the DELETE subscriber operation to all the device groups that contains the deletedsubscriber. This way, the removal of the subscriber is done in a single operation, no need to manually update the device groups to remove the dangling reference anymore.
This PR improves the response when the list of subscribers is empty: now it returns an empty list ([]) instead of null.
This PR also improves the response in case of subscriber deletion failure: now it returns 500 Internal Server Error.